Add fmi{2|3}Functions.c
